Merge all the refactoring for the CheerpJ and JSweet client

+++ desktop/src/main/java/org/vorthmann/zome/render/jogl/JoglRenderingViewer.java
@@ -7,7 +7,7 @@ import java.awt.Component;
 import java.awt.event.MouseEvent;
 import java.awt.image.BufferedImage;
 
-import javax.vecmath.Matrix4d;
+import javax.vecmath.Matrix4f;
 import javax.vecmath.Vector3f;
 
 import com.jogamp.opengl.GL2;
@@ -22,9 +22,9 @@ import com.jogamp.opengl.math.FloatUtil;
 import com.jogamp.opengl.math.Ray;
 import com.jogamp.opengl.util.FPSAnimator;
 import com.jogamp.opengl.util.awt.AWTGLReadBufferUtil;
+import com.vzome.core.construction.Color;
 import com.vzome.core.math.Line;
 import com.vzome.core.math.RealVector;
-import com.vzome.core.model.Color;
 import com.vzome.core.render.RenderedManifestation;
 import com.vzome.core.render.Scene;
 import com.vzome.core.viewing.Lights;
@@ -163,9 +163,9 @@ public class JoglRenderingViewer implements RenderingViewer, GLEventListener
     // These are RenderingViewer methods
     
     @Override
-    public void setViewTransformation( Matrix4d trans )
+    public void setViewTransformation( Matrix4f trans )
     {
-        Matrix4d copy = new Matrix4d();
+        Matrix4f copy = new Matrix4f();
         copy .invert( trans );
         int i = 0;
         // JOGL requires column-major order
